+++ CHANGELOG.md
@@ -9,3 +9,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Remove absolute values from metric reports
 * Support block syntax for `trigger!` to narrow metric changes
+* Add tracking of memory usage of a specific object (`Mnemonic.new() { |c| c.memsize_of(obj) }`)